pyproject.toml
[project]
name = "meu-projeto-tts"
version = "0.1.0"
description = "Converte Markdown para Audiobook usando Kokoro TTS ou XTTS v2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydub>=0.25.1",
    "kokoro>=0.9.2",
    "soundfile>=0.12.0",
    "numpy>=1.21.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
xtts = [
    "tts>=0.22.0",
    "transformers>=4.26.0,<4.35.0",
    "librosa>=0.9.0",
]
edge = [
    "edge-tts>=6.1.0",
]
edge-xtts = [
    "edge-tts>=6.1.0",
    "tts>=0.22.0",
    "transformers>=4.26.0,<4.35.0",
    "librosa>=0.9.0",
]
piper = [
    "piper-tts>=1.2.0",
    "onnxruntime>=1.16.0",
]
dev = [
    "mutagen>=1.46.0",
    "ruff>=0.6.0",
]

[project.scripts]
tts-audiobook = "main:main"

[tool.ruff]
line-length = 88
target-version = "py310"
